Task 121
Troubing bugs resolved / commented.

+++ src/net/sf/memoranda/util/FileStorage.java
@@ -256,8 +256,10 @@ public class FileStorage implements Storage {
         String id = prj.getID();
         File f = new File(JN_DOCPATH + id);
         File[] files = f.listFiles();
-        for (int i = 0; i < files.length; i++)
-            files[i].delete();
+        if(files != null){
+	        for (int i = 0; i < files.length; i++)
+	            files[i].delete();
+        }
         f.delete();
     }
